[Feature #154305252] Restructure components into separate folders
