+++ src/App.css
@@ -1,13 +1,13 @@
 body,
 html {
     height: 100%;
-    background-repeat: no-repeat;
-    background-image: radial-gradient(rgb(247, 159, 44), rgb(34, 180, 158));
+    /* background-repeat: no-repeat;
+    background-image: radial-gradient(rgb(247, 159, 44), rgb(34, 180, 158)); */
     font-size: 16px;
 }
 
 .card-container.card {
-  max-width: 500px;
+  max-width: 450px;
   padding: 40px 40px;
 }
