+++ src/components/listItems/shoppingListItemForm.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import Field from './field.js';
+import Field from '../field.js';
 import validator from 'validator';
 
 class ItemForm extends Component{
