+++ src/components/shoppingList/shoppingList.js
@@ -1,8 +1,7 @@
 /* shopping list component to display shopping lists in a table */
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import TableHeading from './tableHeading';
-import { Table, Glyphicon } from 'react-bootstrap';
+import TableHeading from '../tableHeading';
 import ShoppingListForm from './shoppingListForm';
 
 class ShoppingListDashboard extends Component {
@@ -73,7 +72,7 @@ const ShoppingListTable = (props) => {
         );
     });
     return (
-        <Table> 
+        <table> 
             <thead>
                 <tr>
                     <TableHeading heading="Name"/>
@@ -81,7 +80,7 @@ const ShoppingListTable = (props) => {
                 </tr>
             </thead>
             <tbody>{rows}</tbody>
-        </Table>
+        </table>
     );
 }
 
@@ -107,9 +106,9 @@ class ShoppingListRow extends Component {
                 <td>{shoppingList.name}</td>
                 <td>{shoppingList.due_date}</td>
                 <td colSpan={2}>
-                    <button><Glyphicon glyph="eye-open"/> View</button>
-                    <button onClick={this.onClickUpdate}><Glyphicon glyph="edit"/> Update</button>
-                    <button onClick={this.onClickDelete}><Glyphicon glyph="remove"/> Delete</button>
+                    <button> View</button>
+                    <button onClick={this.onClickUpdate}>Update</button>
+                    <button onClick={this.onClickDelete}>Delete</button>
                 </td>
             </tr>
         );
@@ -162,7 +161,7 @@ class ToggleableShoppingListForm extends Component {
             );
         } else {
             return(
-                <button type="button" onClick={this.handleAddClick}><Glyphicon glyph="plus"/></button>
+                <button type="button" onClick={this.handleAddClick}></button>
             );
         }
     }
